server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
pymysql==1.1.0
cryptography==41.0.4
python-dotenv==1.0.0
